solver/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ortools==9.7.2996
pandas==2.1.3
openpyxl==3.1.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
prometheus-fastapi-instrumentator==6.1.0
httpx==0.25.0
